requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
celery==3.1.19
certifi==2015.11.20.1
Django==1.8.7
django-datatables-view==1.12.1
django-dbarray==0.2
django-tastypie==0.12.2
-e git+git@github.com:bordercore/PyOrgMode.git@d43dac98c938313f029d290155e8c136ec2a537f#egg=PyOrgMode-master
-e hg+https://bordercore@code.google.com/r/bordercore-solrpy/@0c95da857c811e6ebafab06a66e5293da6d8d5e8#egg=solrpy-dev
feedparser==5.1.3
flower==0.8.3
flup==1.0.2
futures==3.0.3
google-api-python-client==1.3.2
httplib2==0.9
jsonfield==1.0.3
kombu==3.0.33
lxml==3.4.2
Markdown==2.6
mutagen==1.28
nose==1.3.7
oauth2client==1.4.6
pipdeptree==0.5.0
psycopg2==2.6
pyasn1==0.1.7
pyasn1-modules==0.0.5
Pygments==2.0.2
python-amazon-product-api==0.2.8
python-dateutil==2.4.1
python-mimeparse==0.1.4
pytz==2015.7
requests==2.5.3
rsa==3.1.4
simplejson==3.6.5
singledispatch==3.4.0.3
six==1.9.0
sqlparse==0.1.15
tornado==4.3
uritemplate==0.6
wsgiref==0.1.2
